+++ app/src/main/java/com/undef/localhandsbrambillafunes/ui/navigation/Navigation.kt
@@ -13,12 +13,12 @@ import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import com.undef.localhandsbrambillafunes.data.local.db.ApplicationDatabase
 import com.undef.localhandsbrambillafunes.data.local.repository.FavoriteRepository
-import com.undef.localhandsbrambillafunes.data.local.viewmodel.ProductViewModel
-import com.undef.localhandsbrambillafunes.data.local.viewmodel.SessionViewModel
-import com.undef.localhandsbrambillafunes.data.local.viewmodel.SessionViewModelFactory
+import com.undef.localhandsbrambillafunes.ui.viewmodel.products.ProductViewModel
+import com.undef.localhandsbrambillafunes.ui.viewmodel.session.SessionViewModel
+import com.undef.localhandsbrambillafunes.ui.viewmodel.session.SessionViewModelFactory
 import com.undef.localhandsbrambillafunes.data.local.repository.UserRepository
-import com.undef.localhandsbrambillafunes.data.local.viewmodel.FavoriteViewModel
-import com.undef.localhandsbrambillafunes.data.local.viewmodel.FavoriteViewModelFactory
+import com.undef.localhandsbrambillafunes.ui.viewmodel.favorites.FavoriteViewModel
+import com.undef.localhandsbrambillafunes.ui.viewmodel.favorites.FavoriteViewModelFactory
 import com.undef.localhandsbrambillafunes.ui.screens.auth.ForgotPasswordScreen
 import com.undef.localhandsbrambillafunes.ui.screens.auth.LoginScreen
 import com.undef.localhandsbrambillafunes.ui.screens.auth.RegisterScreen
